changes made by the boss
